init Response refactor, add Body
